+++ pxdlib/pxdfile.py
@@ -8,13 +8,14 @@ import sqlite3
 from io import UnsupportedOperation
 from collections import namedtuple
 
+from .helpers import SizeHelper
 from .layer import _LAYER_TYPES, Layer
 from .structure import blob, make_blob
 
 guides = namedtuple('guides', ('horizontal', 'vertical'))
 
 
-class PXDFile:
+class PXDFile(SizeHelper):
     def __repr__(self):
         return f"PXDFile({repr(str(self.path))})"